Project/requirements.txt
streamlit==1.27.0
firebase-admin==6.2.0
streamlit-webrtc==0.47.0
numpy==1.24.3
opencv-python==4.8.0.76
Pillow==10.0.1
tensorflow==2.13.0
mtcnn==0.1.1
pandas==2.0.3
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0 